feat(audit): split into scope-grouped audit / audit:environment / audit:app verbs
Mirrors the sync command shape (account → environment → app). Drops the
`--app` option in favour of a positional `<app>` on `audit:app`. `--drift`
stays universal across all three verbs.

Shared query/classify/render lifted into AbstractAuditCommand; the three
concrete commands provide the per-scope row filter and empty-state wording.
